Removed unused table, and moved files.
